final commit

+++ src/stores/materialsStore.js
@@ -1,6 +1,6 @@
 import { defineStore } from 'pinia'
 import  db  from '../firebase/init'
-import {  collection, getDocs , doc, setDoc , updateDoc , deleteDoc, addDoc} from 'firebase/firestore'
+import {  collection, doc, setDoc , updateDoc , deleteDoc, addDoc, onSnapshot } from 'firebase/firestore'
 
 export const useMaterialsStore = defineStore('materials', {
   state: () => ({
@@ -26,9 +26,12 @@ export const useMaterialsStore = defineStore('materials', {
       // {  name: 'Radiator', stock: 2, stockUsed: 0 },
       // {  name: 'Exhaust System', stock: 1, stockUsed: 0 },
     ],
+    materialsUsed: [],
     isDataFetched: false,
     totalStock: 0,
     totalStockUsed: 0,
+    sortBy: '',
+    sortOrder: 'asc',
   }),
   getters: {
     filteredMaterials: (state) => (searchQuery) => {
@@ -36,52 +39,65 @@ export const useMaterialsStore = defineStore('materials', {
         item.name.toLowerCase().includes(searchQuery.toLowerCase())
       )
     },
-    getTotalStock: async () => {
-      let totalStock = 0;
-
-      const querySnapshot = await getDocs(collection(db, "materials"));
-      querySnapshot.forEach((doc) => {
-        totalStock += doc.data().stock;
+    sortedMaterials: (state) => {
+      return [...state.materialsInventory].sort((a, b) => {
+        if (a[state.sortBy] < b[state.sortBy]) {
+          return state.sortOrder === 'asc' ? -1 : 1;
+        }
+        if (a[state.sortBy] > b[state.sortBy]) {
+          return state.sortOrder === 'asc' ? 1 : -1;
+        }
+        return 0;
       });
-
-      return totalStock;
-    },
-    getTotalStockUsed: async () => {
-      let totalStockUsed = 0;
-
-      const querySnapshot = await getDocs(collection(db, "materialUsed"));
-      querySnapshot.forEach((doc) => {
-        totalStockUsed += doc.data().quantityUsed;
-      });
-      console.log('totalStockUsed:', totalStockUsed)
-      return totalStockUsed;
     }
   },
   actions: {
+    changeSort(column) {
+      if (this.sortBy === column) {
+        this.sortOrder = this.sortOrder === 'asc' ? 'desc' : 'asc';
+      } else {
+        this.sortBy = column;
+        this.sortOrder = 'asc';
+      }
+    },
     async getMaterials() {
       if (this.isDataFetched) return;
 
-      const querySnapshot = await getDocs(collection(db, "materials"));
-      querySnapshot.forEach((doc) => {
-        this.materialsInventory.push(doc.data());
+      onSnapshot(collection(db, "materials"), (querySnapshot) => {
+        this.materialsInventory = [];
+        querySnapshot.forEach((doc) => {
+          this.materialsInventory.push(doc.data());
+        });
+
+        this.calculateTotalStock();
+        this.isDataFetched = true;
       });
-      this.calculateTotalStockUsed();
-      this.calculateTotalStock();
-      this.isDataFetched = true;
     },
+
+    async getMaterialsUsed() {
+      onSnapshot(collection(db, "materialUsed"), (querySnapshot) => {
+        this.materialsUsed = [];
+        querySnapshot.forEach((doc) => {
+          this.materialsUsed.push(doc.data());
+        });
+        this.calculateTotalStockUsed();
+      });
+    },
+
     calculateTotalStock() {
       this.totalStock = this.materialsInventory.reduce(
         (total, item) => total + item.stock,
         0
       )
     },
+
     calculateTotalStockUsed() {
-      this.totalStockUsed = this.materialsInventory.reduce(
-        (total, item) => total + item.stockUsed,
+      this.totalStockUsed = this.materialsUsed.reduce(
+        (total, item) => total + item.quantityUsed,
         0
-      )
-      console.log('Total Stock Used:', this.totalStockUsed)
+      );
     },
+
     async addMaterial(material) {
       const newId =
         this.materialsInventory.length > 0
@@ -90,16 +106,14 @@ export const useMaterialsStore = defineStore('materials', {
 
       this.materialsInventory.push({ ...material, id: newId, stockUsed: 0 });
       this.calculateTotalStock();
-      this.calculateTotalStockUsed();
+
 
       const docRef = doc(db, "materials", newId.toString());
       await setDoc(docRef, {
         ...material,
         id: newId,
-        stockUsed: 0,
       });
     },
-
 
     editMaterial(updatedMaterial) {
       const index = this.materialsInventory.findIndex(
@@ -114,44 +128,54 @@ export const useMaterialsStore = defineStore('materials', {
         const docRef = doc(db, "materials", updatedMaterial.id.toString());
         updateDoc(docRef, updatedMaterial).then(() => {
           this.calculateTotalStock();
-          this.calculateTotalStockUsed();
         })
       }
     },
-    async deleteMaterial(materialId) {
-      // Delete from the state
-      this.materialsInventory = this.materialsInventory.filter(
-        (material) => material.id !== materialId
-      );
-      this.calculateTotalStock();
-      this.calculateTotalStockUsed();
 
-      // Delete from Firestore
-      const docRef = doc(db, "materials", materialId.toString());
-      await deleteDoc(docRef);
+    async deleteMaterial(materialId) {
+      const materialToDelete = this.materialsInventory.find(
+        (material) => material.id === materialId
+      );
+
+      if (materialToDelete) {
+        const materialDeletedRef = collection(db, "materialDeleted");
+        await addDoc(materialDeletedRef, {
+          id: materialToDelete.id,
+          name: materialToDelete.name,
+          stock: materialToDelete.stock,
+          dateDeleted: new Date(),
+        });
+
+        this.materialsInventory = this.materialsInventory.filter(
+          (material) => material.id !== materialId
+        );
+
+        this.calculateTotalStock();
+
+        const docRef = doc(db, "materials", materialId.toString());
+        await deleteDoc(docRef);
+      }
     },
+
     async useMaterial(materialId, quantity) {
       const material = this.materialsInventory.find((m) => m.id === materialId);
       if (material && material.stock >= quantity) {
-        // Update material's stock and stockUsed
         material.stock -= quantity;
-        material.stockUsed += quantity;
 
-        // Update the total stock and stock used in the state
         this.calculateTotalStock();
         this.calculateTotalStockUsed();
 
-        // Update the material in Firestore
         const materialRef = doc(db, "materials", materialId.toString());
         await updateDoc(materialRef, {
           stock: material.stock,
-          stockUsed: material.stockUsed,
         });
 
-        // Add a record to the "materialUsed" collection
         const materialUsedRef = collection(db, "materialUsed");
         await addDoc(materialUsedRef, {
+          id: materialId,
+          name: material.name,
           quantityUsed: quantity,
+          dateUsed: new Date(),
         });
       }
     },
